MTB-269-commit-fixed-entity-service-migration
