+++ backend/src/domain/entities/schema/tag.entity.ts
@@ -1,6 +1,6 @@
 import { Entity, Column, ManyToMany } from "typeorm";
 
-import { App, AppVersion } from "@domain/entities";
+import { App } from "@domain/entities";
 
 import { BaseSoftDelete } from "../base";
 
@@ -14,7 +14,4 @@ export class Tag extends BaseSoftDelete {
 
     @ManyToMany(() => App, (app) => app.tags)
     public apps: App[];
-
-    @ManyToMany(() => AppVersion, (appVersion) => appVersion.tags)
-    public appVersions: AppVersion[];
 }
